Added translation of runtime principals and labels.
Key changes are that parameters of classes are now final fields of the translated objects; constructors take initialiazers for these fields. Static methods take the parameters of the class as additional arguments. Cast and instanceofs are translated into static method calls on the class being cast/compared to.

Still need to write a whole lot of test cases, and some more documentation.