First attempt at exposing the jif.lang.Principal class to Jif code. Code compiles.
Subtypes of the interface jif.lang.Principal implicitly have the authority of the principal "this" (the authority is implicit, rather than explicit, to avoid nasty scoping issues, about when and where "this" is allowed to appear). Subtypes of

Subtypes of the interface jif.lang.Principal can implicitly be cast to the primitive type "principal". The null principal is interpreted as a principal who allows everyone to act for it.

Capabilities and Closures are in there as Jif code, but the Runtime interface does not as yet support the creation of Capabilities.

Need to implement a search for the actsfor chains (currently only tests immediate delegation).

Need to start moving some of the principal code (AbstractPrincipal, NativePrincipal, Alice, Bob, etc.) from Java code into Jif code, and add test cases to exercise the new expressiveness.