Rewrote the implementation of AbstractPrincipal and InternalPrincipal in Jif.
Ideally, we should remove rt-src/jif/label/Principal.java, and sig-src/jif/lang/Principal.jif, so that we have only one signature and implementation for Principal. To do this would require bootstrapping of the compiler, since the runtime code (in java) needs to know about jif.lang.Principal, and that must be compiled before the Jif compiler is ready to compile jif/lang/Principal.jif. How do people feel about a "jif-lib-bootstrap.jar", which contains compiled versions of files in lib-src? This would be sufficient to allow us to delete 2 of the three duplicates signatures for the Principal class.