Some test cases for the authority/closure stuff.
Minor changes to the Capability and Closure signatures.

Fixed bugs with substitution of uninterpreted access paths, using actsfor with non-principals, and allowed the type system to reason about equivalence of principals when local variables are initialiazed with a cast expression, e.g. "ExternalPrincipal p = (ExternalPrincipal)this.q", need to be able to reason that p and this.q are equivalent.