Checkpoint: implementing conjunctive and disjunctive principals
Have currently added them to the type system and the runtime system (but not the parser yet).

Needed to add a runtime representation of the top principal (*), to allow the runtime representation of policies such as {Alice: } == {Alice: *}. The other option would be to have explicit representation of policies where the top principal occurs as reader/writer. The type system does ensure that the top principal cannot appear in any "authority" clause, and the top principal never gives its authority to any closures.

Next steps:
- tidy up the integrity labels, taking advantage of the additional flexibility of con/dis-junctive principals to get rid of the "magic semicolon".
- add parsing for con/dis-junctive principals, and maybe some special cases for printing reader/writer policies, so they look like we're used to.
- figure out robust declassification and implement it.