Changed implementation of integrity in Jif.
Confidentiality policies and integrity policies now each form full lattices. Confidentiality policies and integrity policies are not labels; a PairLabel is a pair of a confidentiality policy and integrity policy. Labels now form full lattices.

Some bugs still to iron out: robustness is disabled at the moment, reasonable defaults need to be worked out for signatures, about a dozen test cases need to be examined to find out why they are failing.

Also, signature code needs to be examined, to make sure it makes sense for integrity.

And a whole lot of test cases need to be written.