First take at implementing the endorse operator.
Cleanly factor out the authorization check and the robustness check.

Needed to introduce a confidentiality projection of a label, and an integrity projection of a label, to allow certain constraints to correctly evaluate to true.

AbstractPrincipal seems difficult to implement. I'm not sure yet if it is impossible or just that I haven't found the right labels for things yet...