Changed names of startLabel, entryPC and externalPC to be consistent: pcBound.
Also made the initial pc of a method a special arg label "caller_pc", which is bounded above by the pc bound. The "this" label is bounded above by caller_pc. Using the bottom label as the initial pc of the method is unsound, since it gives too much integrity; using the pc bound as the initial pc is too imprecise in some cases.

The Jif library and signature files now compile; some of the test cases need to be modified to run with the new label checking.