Added explicit java version attribute to javac tasks in build.xml.
Also, jif-base copies the jif.ppg file over into the classes folder so it would be included in the jar for distribution. This was done because split.ppg from swift includes jif.ppg, and it should be found in the classpath. (The current version of split.ppg has an ugly include using ../../../, which would only work if the jif home dir is right next to the swift dir.)