Got past the exception problems.
But I don't have a way to keep the answer to a directory lookup connected
to its (dynamic) label. Both of these need to be updated at each loop iteration.
I can compute the right thing within Jif, but I can't convince the type system
that it is the right thing.

Seem to need a flow-sensitive or SSA-style analysis both to track ordinary
variables and label variables. There is probably a result lurking here.