Comment for potential bug in LabelEnv_c.equivalentAccessPaths. Working on a testcase.
This bug is usually hidden since equivalentAccessPaths is primarily called in
dynamicprincpal actsfor checks.  There, when equivalentAccessPaths returns
false, it falls back on the principal hierarchy, which gives the correct
result.

equivalentAccessPaths is also called from field label checks, but an attempt to
create a testcase for that resulted in an (unrelated?) exception in
JifTypeSystem.processFAP.