Fixed the bug reported by Cormac Flanagan, as exercised by tests/UnsoundLabelInference01.jif. This was done by adding a "LabelsDoubleChecked" goal. This runs the label checker a second time to ensure any labels inferred during the first run are sound to use.
While this approach ensures soundness, it reduces completeness in
principle. Ideally, we would modify the label solver to reason about the
labels of the labels it infers, but it appears that the current solver
design would not know how to adjust a solution that was unsound because
of the labels of inferred labels.