This reverts commit 64e76265c8edaf3f22560b456d6419747b9d2e10 (which made FieldLabelInference and TypeChecked corequisites for an undocumented reason) and makes FieldLabelInference a barrier for LabelChecked.
Making the corresponding change in the master branch does not appear to
cause any regressions in the Jif tests or any of the working Fabric
examples (auction, auth, friendmap, groups, hello, sif-hello, travel),
and causes the "Regression01A.jif Regression01B.jif" test to pass in
this branch, so I am calling this a win. The "Regression01A.jif,
Regression01B.jif" separate-compilation test still needs to be fixed,
however.